+++ monocle/db/queries.py
@@ -72,6 +72,20 @@ public_queries = (
 )
 
 
+def ensure_gte_lte(es, index, repository_fullname, params):
+    if not params.get('gte'):
+        first_created_event = _first_created_event(
+            es, index, repository_fullname, params
+        )
+        if first_created_event:
+            params['gte'] = int(is8601_to_dt(first_created_event).timestamp() * 1000)
+        else:
+            # There is probably nothing in the db that match the query
+            params['gte'] = None
+    if not params.get('lte'):
+        params['lte'] = int(utcnow().timestamp() * 1000)
+
+
 def generate_events_filter(params, qfilter):
     gte = params.get('gte')
     lte = params.get('lte')
@@ -115,7 +129,9 @@ def generate_changes_filter(params, qfilter):
             )
 
 
-def generate_filter(repository_fullname, params):
+def generate_filter(es, index, repository_fullname, params, ensure_time_range=True):
+    if ensure_time_range:
+        ensure_gte_lte(es, index, repository_fullname, params)
     gte = params.get('gte')
     lte = params.get('lte')
     etype = params.get('etype')
@@ -199,7 +215,7 @@ def _scan(es, index, repository_fullname, params):
     body = {
         # "_source": "change_id",
         "_source": params.get('field', []),
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     scanner_params = {'index': index, 'query': body}
     data = scanner(es, **scanner_params)
@@ -215,7 +231,9 @@ def _scan(es, index, repository_fullname, params):
 def _first_created_event(es, index, repository_fullname, params):
     body = {
         "sort": [{"created_at": {"order": "asc"}}],
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(
+            es, index, repository_fullname, params, ensure_time_range=False
+        ),
     }
     data = run_query(es, index, body)
     data = [r['_source'] for r in data['hits']['hits']]
@@ -224,7 +242,7 @@ def _first_created_event(es, index, repository_fullname, params):
 
 
 def count_events(es, index, repository_fullname, params):
-    body = {"query": generate_filter(repository_fullname, params)}
+    body = {"query": generate_filter(es, index, repository_fullname, params)}
     count_params = {'index': index}
     count_params['body'] = body
     res = es.count(**count_params)
@@ -237,7 +255,7 @@ def count_authors(es, index, repository_fullname, params):
             "agg1": {"cardinality": {"field": "author", "precision_threshold": 3000}}
         },
         "size": 0,
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     return data['aggregations']['agg1']['value']
@@ -272,6 +290,7 @@ def interval_to_format(interval):
 
 
 def events_histo(es, index, repository_fullname, params):
+    _filter = generate_filter(es, index, repository_fullname, params)
     duration = (params['lte'] - params['gte']) / 1000
     interval = set_histo_granularity(duration)
     fmt = interval_to_format(interval)
@@ -290,7 +309,7 @@ def events_histo(es, index, repository_fullname, params):
             "avg_count": {"avg_bucket": {"buckets_path": "agg1>_count"}},
         },
         "size": 0,
-        "query": generate_filter(repository_fullname, params),
+        "query": _filter,
     }
     data = run_query(es, index, body)
     return (
@@ -300,6 +319,7 @@ def events_histo(es, index, repository_fullname, params):
 
 
 def authors_histo(es, index, repository_fullname, params):
+    _filter = generate_filter(es, index, repository_fullname, params)
     duration = (params['lte'] - params['gte']) / 1000
     interval = set_histo_granularity(duration)
     fmt = interval_to_format(interval)
@@ -319,7 +339,7 @@ def authors_histo(es, index, repository_fullname, params):
             "avg_count": {"avg_bucket": {"buckets_path": "agg1>_count"}},
         },
         "size": 0,
-        "query": generate_filter(repository_fullname, params),
+        "query": _filter,
     }
     data = run_query(es, index, body)
     res = data["aggregations"]["agg1"]["buckets"]
@@ -340,7 +360,7 @@ def _events_top(es, index, repository_fullname, field, params):
             }
         },
         "size": 0,
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     count_series = [b['doc_count'] for b in data['aggregations']['agg1']['buckets']]
@@ -365,6 +385,7 @@ def repos_top(es, index, repository_fullname, params):
 
 
 def events_top_authors(es, index, repository_fullname, params):
+    params = deepcopy(params)
     return _events_top(es, index, repository_fullname, "author", params)
 
 
@@ -462,7 +483,7 @@ def change_merged_count_by_duration(es, index, repository_fullname, params):
             }
         },
         "size": 0,
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     return data['aggregations']['agg1']['buckets']
@@ -476,7 +497,7 @@ def change_merged_avg_duration(es, index, repository_fullname, params):
         "aggs": {"agg1": {"avg": {"field": "duration"}}},
         "size": 0,
         "docvalue_fields": [{"field": "created_at", "format": "date_time"}],
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     return data['aggregations']['agg1']['value']
@@ -490,7 +511,7 @@ def change_merged_avg_commits(es, index, repository_fullname, params):
         "aggs": {"agg1": {"avg": {"field": "commit_count"}}},
         "size": 0,
         "docvalue_fields": [{"field": "created_at", "format": "date_time"}],
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     return data['aggregations']['agg1']['value']
@@ -797,7 +818,7 @@ def last_changes(es, index, repository_fullname, params):
         "sort": [{"updated_at": {"order": "desc"}}],
         "size": params['size'],
         "from": params['from'],
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     changes = [r['_source'] for r in data['hits']['hits']]
@@ -832,7 +853,7 @@ def oldest_open_changes(es, index, repository_fullname, params):
         "sort": [{"created_at": {"order": "asc"}}],
         "size": params['size'],
         "from": params['from'],
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     changes = [r['_source'] for r in data['hits']['hits']]
@@ -856,7 +877,7 @@ def changes_and_events(es, index, repository_fullname, params):
         "sort": [{"created_at": {"order": "asc"}}],
         "size": params['size'],
         "from": params['from'],
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     changes = [r['_source'] for r in data['hits']['hits']]
@@ -886,7 +907,7 @@ def changes_by_file_map(es, index, repository_fullname, params):
     params['etype'] = ("Change",)
     body = {
         "size": 1000,
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     changes = [r['_source'] for r in data['hits']['hits']]
@@ -903,7 +924,7 @@ def authors_by_file_map(es, index, repository_fullname, params):
     params['etype'] = ("Change",)
     body = {
         "size": 1000,
-        "query": generate_filter(repository_fullname, params),
+        "query": generate_filter(es, index, repository_fullname, params),
     }
     data = run_query(es, index, body)
     changes = [r['_source'] for r in data['hits']['hits']]